isolated AJAX calls into their own function
I took the 2 ajax calls out of the search button click listener function and wrapped them in their own function called 'APIcalls', then invoked APIcalls inside the click listener so the AJAX calls still run on click. Then I set a new click listener on the search history buttons, which also invoked the APIcalls function. The only functional difference between the 2 click listeners is how the city is defined. However, once the user uses the search bar, the list buttons stop responding. This is an issue that must be resolved.